feat: get the facade for vocab cards working
It is all fake though; we have abstracted the data access for now so to not involve any APIs right now.
Idea is to do as much work as possible in the UI, so implementing a backend for this would be easier.
Next step is to building the mechanism to optimally store user's learned words based on their feedback, and present them correct words to guess instead of just random words. Oh and also to make examples work.