feat: Exercises
Initial draft. I think it's better to first implement an Anki clone, and then
enhance it with Duolingo like features, and eventually make it smart with German
specific features. My initial plan of Starting out with German features
implemented on top of Anki+Duolingo (in my mind) isn't proving to be pragmatic.